made utils a subpackage. Distributed the util functions into separate modules. Added logging.py to utils.
