+++ src/component_model/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for component_model."""
